requirements.txt
disnake
colorama
python-dotenv
psutil
humanize